requirements.txt
imutils
numpy
argparse
dlib
opencv-python
mysql-connector-python
datetime
pandas
flask
scipy

